+++ MordoEngine/src/Core/Shader/Shader.cpp
@@ -87,6 +87,11 @@ void Shader::setVec3(const std::string& name, const glm::vec3& value) const
 	glUniform3fv(glGetUniformLocation(ID, name.c_str()), 1, &value[0]);
 }
 
+void Shader::setMat4(const std::string& name, const glm::mat4& mat) const
+{
+	glUniformMatrix4fv(glGetUniformLocation(ID, name.c_str()), 1, GL_FALSE, &mat[0][0]);
+}
+
 Shader::Shader(Shader&& other) noexcept
 	: ID(other.ID)
 {