+++ MordoEngine/src/Core/Texture/Texture.cpp
@@ -1,5 +1,4 @@
 #include "Texture.h"
-#include "../FileSystem/FileSystem.h"
 #include <iostream>
 
 #define STBI_WINDOWS_UTF8
@@ -20,8 +19,7 @@ Texture::Texture(const std::string& texturePath) {
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 	int width, height, nrChannels;
 
-	unsigned char* data = stbi_load(
-		FileSystem::getPath(texturePath).c_str(),
+	unsigned char* data = stbi_load(texturePath.c_str(),
 		&width, &height, &nrChannels, 0);
 	if (data)
 	{