+++ docs/changelog.md
@@ -26,3 +26,15 @@
 - Implemented SquareRenderer to dynamically generate and render textured squares using indexed buffers
 - Replaced TriangleRenderer and Triangle with SquareRenderer
 - Refactored resource managers to use a generic template-based Manager<T> (e.g. textures, shaders)
+
+## 2026-01-29
+- Going 3D!
+- Removed Square and SquareRenderer
+- Added Cube and CubeRenderer for rendering textured 3D cubes with a coordinate system
+- Random cube generation on space key press in CubeRenderer with random positions, rotations, scales
+
+## 2026-01-30
+- Implemented a Camera class for 3D navigation using quaternions
+- Updated CubeRenderer to utilize the Camera for rendering cubes in 3D space
+- Implemented WASD movement and mouse functionality in the cube renderer for camera control
+